Rearrange warnigns
