--- warn/warn_bazel_misc_test.go
@@ -1,43 +1,0 @@
-package warn
-
-import "testing"
-
-func TestConstantGlob(t *testing.T) {
-	checkFindings(t, "constant-glob", `
-cc_library(srcs = glob(["foo.cc"]))
-cc_library(srcs = glob(["*.cc"]))
-cc_library(srcs =
-  ["constant"] + glob([
-    "*.cc",
-    "test.cpp",
-  ])
-)`,
-		[]string{":1: Glob pattern `foo.cc` has no wildcard",
-			":6: Glob pattern `test.cpp` has no wildcard"},
-		scopeEverywhere)
-}
-
-func TestNativeInBuildFiles(t *testing.T) {
-	checkFindingsAndFix(t, "native-build", `
-native.package("foo")
-
-native.cc_library(name = "lib")
-`, `
-package("foo")
-
-cc_library(name = "lib")
-`, []string{
-		`:1: The "native" module shouldn't be used in BUILD files, its members are available as global symbols.`,
-		`:3: The "native" module shouldn't be used in BUILD files, its members are available as global symbols.`,
-	}, scopeBuild)
-}
-
-func TestNativePackage(t *testing.T) {
-	checkFindings(t, "native-package", `
-native.package("foo")
-
-native.cc_library(name = "lib")
-`, []string{
-		`:1: "native.package()" shouldn't be used in .bzl files.`,
-	}, scopeBzl)
-}
